% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{as.data.frame.Spectra}
\alias{as.data.frame.Spectra}
\title{Convert a Spectra Object to a Data Frame}
\usage{
\method{as.data.frame}{Spectra}(x, ...)
}
\arguments{
\item{x}{A \code{Spectra} object.}

\item{...}{Additional arguments (ignored in this method).}
}
\value{
A data frame with two columns: wavelength and intensity.
}
\description{
This function converts a \code{Spectra} object to a data frame.
}
