% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{as.Spectra}
\alias{as.Spectra}
\title{Create a Spectra Object}
\usage{
as.Spectra(wavelength, intensity)
}
\arguments{
\item{wavelength}{A numeric vector of wavelengths.}

\item{intensity}{A numeric vector of intensities corresponding to the wavelengths.}
}
\value{
An object of class \code{Spectra}.
}
\description{
This function creates a \code{Spectra} object, which contains wavelength and intensity data.
}
\examples{
wavelength <- seq(400, 700, length.out = 100)
intensity <- sin(wavelength / 100) + rnorm(100, sd = 0.05)
spectra <- as.Spectra(wavelength, intensity)
print(spectra)
}
