% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline_correct_Spectra}
\alias{baseline_correct_Spectra}
\title{Calculate Baseline-Corrected Spectrum}
\usage{
baseline_correct_Spectra(x)
}
\arguments{
\item{x}{An object of class \code{Spectra}.}
}
\value{
A \code{Spectra} object representing the baseline-corrected spectrum.
}
\description{
This function calculates the baseline-corrected spectrum using the convex hull method.
}
\examples{
wavelength <- seq(400, 700, length.out = 100)
intensity <- sin(wavelength / 100) + rnorm(100, sd = 0.05)
spectra <- as.Spectra(wavelength, intensity)
corrected_spectrum <- baseline_correct_Spectra(spectra)
plot(spectra, corrected_spectrum = corrected_spectrum)
}
