% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{convex_hull}
\alias{convex_hull}
\title{Generic Convex Hull Function}
\usage{
convex_hull(x, ...)
}
\arguments{
\item{x}{The object for which the convex hull should be calculated.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
The output depends on the specific method applied.
}
\description{
This function checks the class of the input and calls the appropriate convex hull calculation function.
}
\examples{
wavelength <- seq(400, 700, length.out = 100)
intensity <- sin(wavelength / 100) + rnorm(100, sd = 0.05)
spectra <- as.Spectra(wavelength, intensity)
hull_points <- convex_hull(spectra)
}
