% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{convex_hull_Spectra}
\alias{convex_hull_Spectra}
\title{Calculate Convex Hull for Spectra}
\usage{
convex_hull_Spectra(x)
}
\arguments{
\item{x}{An object of class \code{Spectra}.}
}
\value{
A matrix of points representing the convex hull.
}
\description{
This function calculates the convex hull for a \code{Spectra} object using the convex hull algorithm.
}
\examples{
wavelength <- seq(400, 700, length.out = 100)
intensity <- sin(wavelength / 100) + rnorm(100, sd = 0.05)
spectra <- as.Spectra(wavelength, intensity)
hull_points <- convex_hull_Spectra(spectra)
plot(spectra, hull_points = hull_points)
}
