% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{import_spectra}
\alias{import_spectra}
\title{Import a Dataset from a CSV File and Create a Spectra Object}
\usage{
import_spectra(
  file_path,
  wavelength_col = "wavelength",
  intensity_col = "intensity"
)
}
\arguments{
\item{file_path}{The path to the CSV file containing the dataset.}

\item{wavelength_col}{The name of the column containing wavelength data. Defaults to "wavelength".}

\item{intensity_col}{The name of the column containing intensity data. Defaults to "intensity".}
}
\value{
An object of class \code{Spectra}.
}
\description{
This function imports a dataset from a CSV file, checks for the required column names,
and then creates a \code{Spectra} object. The user can specify which columns to use for wavelength and intensity.
}
\examples{
file_path <- system.file("extdata", "test_spectra.csv", package = "baselinedetection")
spectra <- import_spectra(file_path)
print(spectra)

}
