% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{plot.Spectra}
\alias{plot.Spectra}
\title{Plot a Spectra Object}
\usage{
\method{plot}{Spectra}(x, hull_points = NULL, corrected_spectrum = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{Spectra}.}

\item{hull_points}{A matrix of convex hull points to overlay on the plot (optional).}

\item{corrected_spectrum}{A \code{Spectra} object representing the baseline-corrected spectrum (optional).}

\item{...}{Additional graphical parameters to be passed to \code{plot}.}
}
\description{
This function plots the \code{Spectra} object, with optional overlays for convex hull and corrected spectrum.
}
\examples{
wavelength <- seq(400, 700, length.out = 100)
intensity <- sin(wavelength / 100) + rnorm(100, sd = 0.05)
spectra <- as.Spectra(wavelength, intensity)
plot(spectra)
}
