% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{plot_spectra_corrected}
\alias{plot_spectra_corrected}
\title{Plot Corrected Spectra Object}
\usage{
plot_spectra_corrected(corrected_spectrum, ...)
}
\arguments{
\item{corrected_spectrum}{A \code{Spectra} object representing the baseline-corrected spectrum.}

\item{...}{Additional graphical parameters to be passed to \code{plot}.}
}
\description{
This function plots the baseline-corrected \code{Spectra} object.
}
\examples{
wavelength <- seq(400, 700, length.out = 100)
intensity <- sin(wavelength / 100) + rnorm(100, sd = 0.05)
spectra <- as.Spectra(wavelength, intensity)
corrected_spectrum <- baseline_correct_Spectra(spectra)
plot_spectra_corrected(corrected_spectrum)
}
