% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{print.Spectra}
\alias{print.Spectra}
\title{Print a Spectra Object}
\usage{
\method{print}{Spectra}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Spectra}.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
This function prints the summary information of a \code{Spectra} object.
}
\examples{
wavelength <- seq(400, 700, length.out = 100)
intensity <- sin(wavelength / 100) + rnorm(100, sd = 0.05)
spectra <- as.Spectra(wavelength, intensity)
print(spectra)
}
